refactor: clean explorer utilities
